BUG LambdaParserTest: Workaround for missing URL reference to rt.jar in AppClassLoader during "mvn install"
Details: Guava's ClassPath does not consider java.util.Collections a (top-level) class during "mvn install" which is why testBiConsumer() and testTypedBiConsumer() failed with exceptions during the unit tests. It seems that the AppClassLoader does not contain an URL for rt.jar as it does when the unit tests are executed regularly. As a workaround, we import the class java.util.Collection (which can be loaded nevertheless) and use the unqualified class name "Collections" instead of the fully qualified class name "java.util.Collections" in the expressions.