ThenRule does not need to restore the parser state in case of an exception
Rationale: It is the responsibility of the ones that are parsing a rule and require that the parser is in the previous state in case of an exception. Most rules don't care because with an exception their work is done. Only a few, like the OrRule and the RepetitionRule, which can recover from exceptions, need to call this method.